refactor: compute graphviz layouts independently of tikz export
